refactor: 패키지명과 모듈명을 짧은 이름으로 변경
